uar-949 rename fetch overseas entity email util
